requirements.txt
black==24.4.2
Django==5.0.7
django-ninja==1.2.1
environs==11.0.0
faker==26.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.32.3
